[#115362749] do same as other branch

+++ ambassadorsdk/src/main/java/com/ambassador/ambassadorsdk/internal/ConversionDBHelper.java
@@ -108,7 +108,7 @@ public class ConversionDBHelper extends SQLiteOpenHelper {
                         )
                 )
                 .setRevenue(
-                        cursor.getInt(
+                        cursor.getFloat(
                                 cursor.getColumnIndex(ConversionSQLStrings.ConversionSQLEntry.MBSY_REVENUE)
                         )
                 )
